BAU: Move variables to common test variables
Tests were referencing these fields which were in another test (the StartHandlerTest), so moving these to a more appropriate shared place
